* Drastically cut down AbstractItem to reduce responsibilities, making * it only responsible for drawing a singular item.
* Fixed all compile time errors in CalendarScene related to incorrect
* data type assignments.

* Changed std::unique_ptr in ApplicationContext to QSharedPointers to
* improve the "Qt'ness" of this god forsaken project.

* Fixed minor compile time errors in CalendarItem.

* Swapped quint16 to quint32 in CustomDateTime for future proofing (in
* case large values will be needed).

* Other minor changes.

+++ include/customdatetime.h
@@ -10,36 +10,36 @@ class CalendarSystem;
 class CustomDateTime
 {
 private:
-    quint16 m_day;
+    quint32 m_day;
     quint32 m_dayAbsolute;
-    quint16 m_month;
+    quint32 m_month;
     quint32 m_year;
-    quint16 m_hour;
-    quint16 m_minute;
-    quint16 m_second;
+    quint32 m_hour;
+    quint32 m_minute;
+    quint32 m_second;
 
 public:
     CustomDateTime();
-    CustomDateTime(quint16 day, quint16 month, quint32 year,
-                   quint16 hour = 0, quint16 minute = 0, quint16 second = 0);
+    CustomDateTime(quint32 day, quint32 month, quint32 year,
+                   quint32 hour = 0, quint32 minute = 0, quint32 second = 0);
 
     // Геттеры
-    quint16 day() const { return m_day; }
+    quint32 day() const { return m_day; }
     quint32 dayAbsolute() const { return m_dayAbsolute; }
-    quint16 month() const { return m_month; }
+    quint32 month() const { return m_month; }
     quint32 year() const { return m_year; }
-    quint16 hour() const { return m_hour; }
-    quint16 minute() const { return m_minute; }
-    quint16 second() const { return m_second; }
+    quint32 hour() const { return m_hour; }
+    quint32 minute() const { return m_minute; }
+    quint32 second() const { return m_second; }
 
     // Сеттеры с валидацией
-    void setDate(quint16 day, quint16 month, quint32 year);
-    void setDay(quint16 day);
-    void setMonth(quint16 month);
-    void setYear(quint16 year);
-    void setTime(quint16 hour, quint16 minute, quint16 second = 0);
-    void setDateTime(quint16 day, quint16 month, quint32 year,
-                     quint16 hour, quint16 minute, quint16 second = 0);
+    void setDate(quint32 day, quint32 month, quint32 year);
+    void setDay(quint32 day);
+    void setMonth(quint32 month);
+    void setYear(quint32 year);
+    void setTime(quint32 hour, quint32 minute, quint32 second = 0);
+    void setDateTime(quint32 day, quint32 month, quint32 year,
+                     quint32 hour, quint32 minute, quint32 second = 0);
 
     // Валидация
     bool isValid() const;
